requirements.txt
aiogram==3.21.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic-settings==2.10.1
